Moving bfs depth unit tests to search utils tests file.
